+++ module/Application/src/Controller/LabelController.php
@@ -7,14 +7,17 @@ use Zend\Form\Annotation\AnnotationBuilder;
 use Doctrine\ORM\EntityManager;
 use DoctrineModule\Stdlib\Hydrator\DoctrineObject as DoctrineHydrator;
 use Application\Entity\Label;
+use Application\Service\ActivityStreamLogger;
+use Application\Form\ConfirmationForm;
 
 class LabelController extends AbstractActionController
 {
     private $em;
     
-    public function __construct(EntityManager $em)
+    public function __construct(EntityManager $em, ActivityStreamLogger $asl)
     {
         $this->em = $em;
+        $this->asl = $asl;
     }
 
     public function indexAction()
@@ -42,8 +45,7 @@ class LabelController extends AbstractActionController
             $form->setData($request->getPost());
             if ($form->isValid()){  
                 $this->em->persist($label); 
-                $this->em->flush();
-                // TODO add activity stream logging        
+                $this->em->flush();                    
                 return $this->redirect()->toRoute('labels');
             }
         }
@@ -61,10 +63,36 @@ class LabelController extends AbstractActionController
             return $this->redirect()->toRoute('labels');
         }
         $label = $this->em->getRepository(Label::class)->find($id);
-        $this->em->remove($label);
-        $this->em->flush();
-        // TODO add activity stream logging        
-        return $this->redirect()->toRoute('labels');
+        if (!$label) {
+            return $this->redirect()->toRoute('labels');
+        } 
+
+        $builder = new AnnotationBuilder();
+        $form = $builder->createForm(new ConfirmationForm());
+        $form->setAttribute('action', $this->url()->fromRoute('labels', array('action' => 'delete', 'id' => $id)));
+        $form->get('cancelTo')->setValue($this->url()->fromRoute('labels'));
+        
+        $request = $this->getRequest();
+        if ($request->isPost()){
+            $form->setData($request->getPost());
+            if ($form->isValid()) { 
+                $data = $form->getData();
+                if ($data['confirm'] == 1) {
+                    $this->em->remove($label);
+                    $this->em->flush();                    
+                } 
+            }
+            return $this->redirect()->toRoute('labels');
+        } 
+
+        $viewModel = new ViewModel(array(
+            'form' => $form,
+            'entityType' => 'label',
+            'entityDescriptor' => $label->getName(),            
+        ));
+        $viewModel->setTerminal($request->isXmlHttpRequest());
+        $viewModel->setTemplate('application/common/confirm.phtml');
+        return $viewModel;
     }
     
 }
